vim: Remove "~" as a leader key
Turns out that the ~ is not a good key to be the leader key even in
abnt2 keyboards. Altough it is in a good possition, it is often
configured as a dead key so it can be usefull as a accent. This has the
effect of having to type it twice to trigger the leader key.