bash: Change TERM to support a 256 color tmux
According to the tmux documentation the TERM variable has to be set to
screen-256color in order to have a color terminal in tmux.

Indeed, using xterm-256color does not work. For instance, when running
"make menuconfig" with that configuration, the colors appear mixed up
and in the wrong places.